+++ stories/Autocomplete.stories.js
@@ -1,16 +1,14 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import Autocomplete from '../src/Autocomplete';
+import { text, withKnobs } from '@storybook/addon-knobs';
 
-const stories = storiesOf('javascript/Autocomplete', module);
+const stories = storiesOf('Javascript|Autocomplete', module);
+stories.addDecorator(withKnobs);
 
-stories.addParameters({
-  info: {
-    text: `Add an autocomplete dropdown below your input 
-    to suggest possible values in your form. You can 
-    populate the list of autocomplete options dynamically as well.`
-  }
-});
+const notes = `Add an autocomplete dropdown below your input 
+  to suggest possible values in your form. You can 
+  populate the list of autocomplete options dynamically as well.`
 
 stories.add('Default', () => (
   <Autocomplete
@@ -23,7 +21,7 @@ stories.add('Default', () => (
     }}
     placeholder="Insert here"
   />
-));
+), { notes });
 
 stories.add('With icon', () => (
   <Autocomplete
@@ -35,6 +33,6 @@ stories.add('With icon', () => (
       }
     }}
     placeholder="Insert here"
-    icon="textsms"
+    icon={text('icon', 'textsms')}
   />
-));
+), { notes });