+++ stories/Collection.stories.js
@@ -5,7 +5,7 @@ import Row from '../src/Row';
 import Collection from '../src/Collection';
 import CollectionItem from '../src/CollectionItem';
 
-const stories = storiesOf('components/Collection', module);
+const stories = storiesOf('Components|Collection', module);
 
 stories.addParameters({
   info: {
@@ -13,7 +13,13 @@ stories.addParameters({
   }
 });
 
-stories.addDecorator(story => <Row><Col m={6} s={12}>{story()}</Col></Row>)
+stories.addDecorator(story => (
+  <Row>
+    <Col m={6} s={12}>
+      {story()}
+    </Col>
+  </Row>
+));
 
 stories.add('Basic', () => (
   <Collection>
@@ -26,15 +32,17 @@ stories.add('Basic', () => (
 
 stories.add('Links', () => (
   <Collection>
-    <CollectionItem href='#'>Alvin</CollectionItem>
-    <CollectionItem href='#' active>Alvin</CollectionItem>
-    <CollectionItem href='#'>Alvin</CollectionItem>
-    <CollectionItem href='#'>Alvin</CollectionItem>
+    <CollectionItem href="#">Alvin</CollectionItem>
+    <CollectionItem href="#" active>
+      Alvin
+    </CollectionItem>
+    <CollectionItem href="#">Alvin</CollectionItem>
+    <CollectionItem href="#">Alvin</CollectionItem>
   </Collection>
 ));
 
 stories.add('Header', () => (
-  <Collection header='First Names'>
+  <Collection header="First Names">
     <CollectionItem>Alvin</CollectionItem>
     <CollectionItem>Alvin</CollectionItem>
     <CollectionItem>Alvin</CollectionItem>